chg: usr: enable Regular Expression searches with -S option
files affected:
    - CoM/bin/CoM

chg: usr: changed "Procedures:" to "Details:" to be more generic

Sometimes we just want to add a comment to the change record describing
why we did something, rather than show one or more diff files of the
changes made to the file(s).  This way the contents of this section
can really be anything the user wants to say as to what the change was.

chg: dev: minor refactoring !cosmetic